app/backend/requirements.txt
fastapi==0.110.1
uvicorn==0.25.0
gunicorn==21.2.0
boto3>=1.34.129
requests-oauthlib>=1.2.0,<2.0.0
cryptography>=42.0.8
python-dotenv>=1.0.1
pymongo==4.9.1
pydantic>=2.6.4
email-validator>=2.2.0
pyjwt>=2.10.1
bcrypt==4.1.3
passlib>=1.7.4
tzdata>=2024.2
motor==3.6.0
certifi>=2024.8.30
pytest>=8.0.0
pytest-asyncio>=0.21.0
black>=24.1.1
isort>=5.13.2
flake8>=7.0.0
mypy>=1.8.0
types-aiofiles>=25.1.0
types-requests>=2.31.0
types-urllib3>=1.26.0
types-python-dateutil>=2.8.0
python-jose>=3.3.0
requests>=2.31.0
httpx>=0.27.0
pandas>=2.2.0
numpy>=1.26.0
python-multipart>=0.0.9
jq>=1.6.0
typer>=0.9.0
playwright>=1.40.0
selenium>=4.15.0
undetected-chromedriver>=3.5.4
fake-useragent>=1.4.0
python-facebook-api>=0.20.1
facebook-sdk>=3.1.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
pillow>=10.0.0
aiofiles>=23.2.0
asyncio-throttle>=1.0.2
supabase>=2.23.0
